change!: rename `Cache` to `Stack` because it's more fitting.
