+++ gix-worktree/tests/worktree/mod.rs
@@ -1,6 +1,7 @@
-mod cache;
-
 use gix_hash::ObjectId;
+
+mod stack;
+
 pub type Result<T = ()> = std::result::Result<T, Box<dyn std::error::Error>>;
 
 pub fn hex_to_id(hex: &str) -> ObjectId {
