+++ gix-worktree/tests/worktree/stack/mod.rs
